Start moving examples to doc
