+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+DemoCards = "311a05b2-6137-4a5a-b473-18580a3d38b5"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 JSServe = "824d6782-a2ef-11e9-3a09-e5662e0c26f9"
