+++ docs/examples/index.md
@@ -1,0 +1,9 @@
+# [Examples](@id examples)
+
+This section contains thorough examples for using TestParticle.jl.
+
+---
+
+{{{democards}}}
+
+---
