+++ docs/examples/single_particle/demo_ExB_drift.jl
@@ -1,6 +1,17 @@
-# E×B drift
-# More theoretical details can be found in Introduction to Plasma Physics and Controlled 
-# Fusion, F. F. Chen and Computational Plasma Physics, Toshi Tajima.
+# ---
+# title: E×B drift
+# id: demo_ExB
+# date: 2023-04-19
+# author: "[Tiancheng Liu](https://github.com/TCLiuu); [Hongyang Zhou](https://github.com/henry2004y)"
+# julia: 1.9.0
+# description: Simple ExB drift demonstration using Makie
+# ---
+
+# This example demonstrates a single proton motion under uniform E and B fields.
+# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion by F. F. Chen and Computational Plasma Physics, Toshi Tajima.
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using TestParticle: get_gc
@@ -8,7 +19,7 @@ using TestParticleMakie
 using OrdinaryDiffEq
 using StaticArrays
 using LinearAlgebra
-using GLMakie
+import WGLMakie as WM
 
 function uniform_B(x)
     return SA[0, 0, 1e-8]
@@ -18,7 +29,7 @@ function uniform_E(x)
     return SA[1e-9, 0, 0]
 end
 
-# trace the orbit of the guiding center
+## trace the orbit of the guiding center
 function trace_gc!(dx, x, p, t)
     _, _, E, B, sol = p
     xu = sol(t)
@@ -33,7 +44,7 @@ x0 = [1.0, 0, 0]
 v0 = [0.0, 1.0, 0.1]
 stateinit = [x0..., v0...]
 tspan = (0, 20)
-# E×B drift
+## E×B drift
 param = prepare(uniform_E, uniform_B, species=Proton)
 prob = ODEProblem(trace!, stateinit, tspan, param)
 sol = solve(prob, Tsit5(); save_idxs=[1,2,3,4,5,6])
@@ -44,5 +55,5 @@ prob_gc = ODEProblem(trace_gc!, gc_x0, tspan, (param..., sol))
 sol_gc = solve(prob_gc, Tsit5(); save_idxs=[1,2,3])
 
 gc_analytic = Tuple(xu -> getindex(sol_gc(xu[7]), i) for i = 1:3)
-# numeric result and analytic result
+## numeric result and analytic result
 orbit(sol, vars=[(1, 2, 3), gc, gc_analytic])
