+++ docs/examples/single_particle/demo_FLR.jl
@@ -1,6 +1,16 @@
-# Finite Larmor radius effect
-# More theoretical details can be found in Introduction to Plasma Physics and Controlled 
-# Fusion, F. F. Chen.
+# ---
+# title: Finite-Larmor-Radius effect
+# id: demo_FLR
+# date: 2023-04-19
+# author: "[Tiancheng Liu](https://github.com/TCLiuu), [Hongyang Zhou](https://github.com/henry2004y)"
+# julia: 1.9.0
+# description: Finite Larmor radius effect demonstration using Makie
+# ---
+
+# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion by F. F. Chen.
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using TestParticle: get_gc
@@ -8,10 +18,10 @@ using TestParticleMakie
 using OrdinaryDiffEq
 using StaticArrays
 using LinearAlgebra
-using GLMakie
+import WGLMakie as WM
 using Tensors: laplace
 import Tensors: Vec as Vec3
-# using SpecialFunctions
+## using SpecialFunctions
 
 function uniform_B(x)
     return SA[0, 0, 1e-8]
@@ -21,7 +31,7 @@ function nonuniform_E(x)
     return SA[1e-9*cos(0.3*x[1]), 0, 0]
 end
 
-# trace the orbit of the guiding center
+## trace the orbit of the guiding center
 function trace_gc!(dx, x, p, t)
     q, m, E, B, sol = p
     xu = sol(t)
@@ -32,11 +42,11 @@ function trace_gc!(dx, x, p, t)
     v_perp = xu[4:6] - v_par
     r4 = (m*norm(v_perp)/q/norm(Bv))^2/4
     EB(x) = (E(x)×B(x))/norm(B(x))^2
-    # dx[1:3] = EB(xp) + v_par
+    ## dx[1:3] = EB(xp) + v_par
     dx[1:3] = EB(x) + r4*laplace.(EB, Vec3(x...)) + v_par
 
-    # more accurate
-    # dx[1:3] = besselj0(0.3*m*norm(v_perp)/q/norm(Bv))*EB(x) + v_par
+    ## more accurate
+    ## dx[1:3] = besselj0(0.3*m*norm(v_perp)/q/norm(Bv))*EB(x) + v_par
 end
 
 x0 = [1.0, 0, 0]
@@ -53,5 +63,5 @@ prob_gc = ODEProblem(trace_gc!, gc_x0, tspan, (param..., sol))
 sol_gc = solve(prob_gc, Tsit5(); save_idxs=[1,2,3])
 
 gc_analytic = Tuple(xu -> getindex(sol_gc(xu[7]), i) for i = 1:3)
-# numeric result and analytic result
+## numeric result and analytic result
 orbit(sol, vars=[(1, 2, 3), gc, gc_analytic])
