+++ docs/examples/single_particle/demo_curvature_B.jl
@@ -1,6 +1,18 @@
-# Magnetic field curvature drift
-# More theoretical details can be found in Introduction to Plasma Physics and Controlled 
-# Fusion, F. F. Chen and Computational Plasma Physics, Toshi Tajima.
+# ---
+# title: Curl-B drift
+# id: demo_curlB
+# date: 2023-04-19
+# author: "[Tiancheng Liu](https://github.com/TCLiuu); [Hongyang Zhou](https://github.com/henry2004y)"
+# julia: 1.9.0
+# description: Simple magnetic field curvature drift demonstration using Makie
+# ---
+
+# This example demonstrates a single proton motion under a non-uniform B field with gradient and curvature.
+# Similar to the magnetic field gradient drift, analytic calculation should include both of the gradient drift and the curvature drift.
+# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion, F. F. Chen and Computational Plasma Physics by Toshi Tajima.
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using TestParticle: get_gc
@@ -8,12 +20,12 @@ using TestParticleMakie
 using OrdinaryDiffEq
 using StaticArrays
 using LinearAlgebra
-using GLMakie
 using ForwardDiff: gradient, jacobian
+import WGLMakie as WM
 
 function curved_B(x)
-    # satisify ∇⋅B=0
-    # B_θ = 1/r => ∂B_θ/∂θ = 0
+    ## satisify ∇⋅B=0
+    ## B_θ = 1/r => ∂B_θ/∂θ = 0
     θ = atan(x[3]/(x[1]+3))
     r = hypot(x[1]+3, x[3])
     return SA[-1e-7*sin(θ)/r, 0, 1e-7*cos(θ)/r]
@@ -25,7 +37,7 @@ end
 
 abs_B(x) = norm(curved_B(x))  # |B|
 
-# trace the orbit of the guiding center
+## trace the orbit of the guiding center
 function trace_gc!(dx, x, p, t)
     q, m, E, B, sol = p
     xu = sol(t)
@@ -36,8 +48,8 @@ function trace_gc!(dx, x, p, t)
     v_perp = xu[4:6] - v_par
     Ω = q*norm(Bv)/m
     κ = jacobian(B, x)*Bv  # B⋅∇B
-    # v⟂^2*(B×∇|B|)/(2*Ω*B^2) + v∥^2*(B×(B⋅∇B))/(Ω*B^3) + (E×B)/B^2 + v∥
-    dx[1:3] = norm(v_perp)^2*(Bv×gradient_B)/(2*Ω*norm(Bv)^2) + 
+    ## v⟂^2*(B×∇|B|)/(2*Ω*B^2) + v∥^2*(B×(B⋅∇B))/(Ω*B^3) + (E×B)/B^2 + v∥
+    dx[1:3] = norm(v_perp)^2*(Bv×gradient_B)/(2*Ω*norm(Bv)^2) +
                 norm(v_par)^2*(Bv×κ)/Ω/norm(Bv)^3 + (E(x)×Bv)/norm(Bv)^2 + v_par
 end
 
@@ -45,7 +57,7 @@ x0 = [1.0, 0, 0]
 v0 = [0.0, 1.0, 0.1]
 stateinit = [x0..., v0...]
 tspan = (0, 40)
-# E×B drift
+## E×B drift
 param = prepare(zero_E, curved_B, species=Proton)
 prob = ODEProblem(trace!, stateinit, tspan, param)
 sol = solve(prob, Tsit5(); save_idxs=[1,2,3,4,5,6])
@@ -56,7 +68,5 @@ prob_gc = ODEProblem(trace_gc!, gc_x0, tspan, (param..., sol))
 sol_gc = solve(prob_gc, Tsit5(); save_idxs=[1,2,3])
 
 gc_analytic = Tuple(xu -> getindex(sol_gc(xu[7]), i) for i = 1:3)
-# numeric result and analytic result
-# similar to the magnetic field gradient drift
-# analytic calculation should include both of the gradient drift and the curvature drift
+## numeric result and analytic result
 orbit(sol, vars=[(1, 2, 3), gc, gc_analytic])
