+++ docs/examples/single_particle/demo_gradient_B.jl
@@ -1,6 +1,20 @@
-# Magnetic field gradient drift
-# More theoretical details can be found in Introduction to Plasma Physics and Controlled 
-# Fusion, F. F. Chen.
+# ---
+# title: Grad-B drift
+# id: demo_gradB
+# date: 2023-04-19
+# author: "[Tiancheng Liu](https://github.com/TCLiuu); [Hongyang Zhou](https://github.com/henry2004y)"
+# julia: 1.9.0
+# description: Simple magnetic field gradient drift demonstration using Makie
+# ---
+
+# This example demonstrates a single proton motion under a non-uniform B field with gradient ∇B ⊥ B.
+# The orbit of guiding center includes some high order terms, it is different from the
+# formula of magnetic field gradient drift of some textbooks which just preserves the first
+# order term.
+# More theoretical details can be found in Introduction to Plasma Physics and Controlled Fusion by F. F. Chen, and Fundamentals of Plasma Physics by Paul Bellan.
+
+using JSServe: Page # hide
+Page(exportable=true, offline=true) # hide
 
 using TestParticle
 using TestParticle: get_gc
@@ -8,8 +22,8 @@ using TestParticleMakie
 using OrdinaryDiffEq
 using StaticArrays
 using LinearAlgebra
-using GLMakie
 using ForwardDiff: gradient
+import WGLMakie as WM
 
 function grad_B(x)
     return SA[0, 0, 1e-8+1e-9 *x[2]]
@@ -21,7 +35,7 @@ end
 
 abs_B(x) = norm(grad_B(x))
 
-# trace the orbit of the guiding center
+## trace the orbit of the guiding center
 function trace_gc!(dx, x, p, t)
     q, m, E, B, sol = p
     xu = sol(t)
@@ -47,8 +61,5 @@ prob_gc = ODEProblem(trace_gc!, gc_x0, tspan, (param..., sol))
 sol_gc = solve(prob_gc, Tsit5(); save_idxs=[1,2,3])
 
 gc_analytic = Tuple(xu -> getindex(sol_gc(xu[7]), i) for i = 1:3)
-# numeric result and analytic result
-# The orbit of guiding center includes some high order terms, it is different from the
-# formula of magnetic field gradient drift of some textbooks which just preserves the first
-# order term.
+## numeric result and analytic result
 orbit(sol, vars=[(1, 2, 3), gc, gc_analytic])
