+++ src/utility/constants.jl
@@ -5,5 +5,6 @@ const qᵢ = 1.60217662e-19   # proton mass, [C]
 const mᵢ = 1.673557546e-27  # proton mass, [kg]
 const c  = 299792458.       # speed of light, [m/s]
 const μ₀ = 4π*1e-7          # Vacuum permeability, [H/m]
+const kB = 1.38064852e-23   # Boltzmann constant, [m²kg/(s²K)]
 const Rₑ = 6371e3           # Earth radius, [m]
 const BMoment_Earth = [0.0, 0.0, 7.94e22] # [V*s/(A*m)]
